requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
pyautogen==0.2.18
duckduckgo-search==4.1.1
pydantic==2.5.3
msgpack==1.0.7
openai==1.12.0
httpx==0.25.2